pyproject.toml
[tool.poetry]
name = "dynamite-dogs"
version = "0.1.0"
description = "demo project"
authors = ["datateam"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
llama-index = "^0.10.25"
lancedb = "0.5.5"
llama-index-vector-stores-lancedb = "^0.1.3"
python-dotenv = "^1.0.1"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
gradio = "^4.24.0"
llama-index-llms-anthropic = "^0.1.7"
llama-index-postprocessor-colbert-rerank = "^0.1.1"
llama-index-embeddings-ollama = "^0.1.2"
llama-index-packs-code-hierarchy = "^0.1.2"
llama-index-readers-github = "^0.1.7"
langfuse = "^2.21.1"
llama-index-callbacks-langfuse = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
